+++ BudgetApi/WebApplication1/Model/Goal.cs
@@ -5,7 +5,7 @@
         public int UserID { get; set; }
         public int GoalID { get; set; }
         public string GoalName { get; set;}
-        public int GoalAmount { get; set; }
-        public int GoalContribution { get; set; }
+        public decimal GoalAmount { get; set; }
+        public decimal GoalContribution { get; set; }
     }
 }
